Revert "optimze"
This reverts commit d8e915d1c949dfdaab2bab29b83e12744ddf6be0.

+++ SRManager.h
@@ -45,7 +45,7 @@ class SRManager
 
     */
 
-  float getDistance(int index, int type=CM);
+  int getDistance(int index, int type=CM);
 
      /*!
     @method
